producer/requirements.txt
kafka-python==2.0.2
faker==11.3.0
python-dateutil==2.8.2
six==1.16.0
textblob==0.17.1
nltk==3.6.7 